backend/requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.47
Flask-Migrate==3.1.0
Flask-CORS==3.0.10
psycopg2-binary==2.9.3
requests==2.28
werkzeug==2.1.2
